# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SCP (Server Control Panel) REST API",
    "description": "The version number is updated with each SCP release, regardless of whether the API has changed. Breaking changes are handled by versioning individual endpoints in the request path.",
    "version": "2026.0218.151350"
  },
...
  "servers": [
    {
      "url": "/scp-core"
    }
  ],
  "paths": {
...
    "/api/v1/servers": {
      "get": {
        "summary": "Get servers",
        "responses": {
          "200": {
            "description": "Servers successfully loaded.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServerListMinimal"
                  }
                }
              },
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServerListMinimal"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "ip",
            "in": "query",
            "description": "Filter by ip",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 0,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Filter by server name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 0,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Search ignoring case within the property name, nickname, or ipv4Addresses",
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": ""
      }
    },
    "/api/v1/servers/{serverId}": {
      "patch": {
        "summary": "Start - stop server or update attributes like hostname, nickname, uefi, bootorder, ...",
        "responses": {
          "200": {
            "description": "Server patched successfully."
          },
          "202": {
            "description": "Server patch successfully started.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskInfo"
                }
              },
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskInfo"
                }
              }
            }
          },
          "503": {
            "description": "Node is in maintenance mode.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              },
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/merge-patch+json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/ServerStatePatch"
                  },
                  {
                    "$ref": "#/components/schemas/ServerAutostartPatch"
                  },
                  {
                    "$ref": "#/components/schemas/ServerBootorderPatch"
                  },
                  {
                    "$ref": "#/components/schemas/ServerOsOptimizationPatch"
                  },
                  {
                    "$ref": "#/components/schemas/ServerCpuTopologyPatch"
                  },
                  {
                    "$ref": "#/components/schemas/ServerUEFIPatch"
                  },
                  {
                    "$ref": "#/components/schemas/ServerHostnamePatch"
                  },
                  {
                    "$ref": "#/components/schemas/ServerNicknamePatch"
                  },
                  {
                    "$ref": "#/components/schemas/ServerKeyboardLayoutPatch"
                  },
                  {
                    "$ref": "#/components/schemas/ServerSetRootPasswordPatch"
                  }
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "name": "serverId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "stateOption",
            "in": "query",
            "description": "Only applicable if ServerStatePatch is used. Valid values for state ON: POWERCYCLE, RESET. Valid values for state OFF: POWEROFF.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": "Only one attribute at a time."
      },
      "get": {
        "summary": "Get one server",
        "responses": {
          "200": {
            "description": "Server successfully returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Server"
                }
              },
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/Server"
                }
              }
            }
          },
          "404": {
            "description": "Server not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              },
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "serverId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "loadServerLiveInfo",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "description": ""
      }
    },
...
  },
  "components": {
    "schemas": {
      "OsOptimization": {
        "type": "string",
        "enum": [
          "LINUX",
          "WINDOWS",
          "BSD",
          "LINUX_LEGACY",
          "UNKNOWN"
        ]
      },
...
      "Server": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "hostname": {
            "type": "string",
            "nullable": true
          },
          "nickname": {
            "type": "string",
            "nullable": true
          },
          "disabled": {
            "type": "boolean"
          },
          "template": {
            "type": "object",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ServerTemplateMinimal"
              }
            ]
          },
          "serverLiveInfo": {
            "type": "object",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ServerInfo"
              }
            ]
          },
          "ipv4Addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IPv4AddressMinimal"
            }
          },
          "ipv6Addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IPv6AddressMinimal"
            }
          },
          "site": {
            "$ref": "#/components/schemas/Site"
          },
          "snapshotCount": {
            "type": "integer",
            "format": "int32"
          },
          "maxCpuCount": {
            "type": "integer",
            "format": "int32"
          },
          "disksAvailableSpaceInMiB": {
            "type": "integer",
            "format": "int64"
          },
          "rescueSystemActive": {
            "type": "boolean"
          },
          "snapshotAllowed": {
            "type": "boolean"
          },
          "architecture": {
            "$ref": "#/components/schemas/Architecture"
          }
        }
      },
      "ResponseError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "TaskInfo": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TaskState"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2022-03-10T16:15:50Z"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2022-03-10T16:15:50Z"
          },
          "executingUser": {
            "$ref": "#/components/schemas/UserMinimal"
          },
          "taskProgress": {
            "$ref": "#/components/schemas/TaskProgress"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "onRollback": {
            "type": "boolean"
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskInfoStep"
            }
          },
          "result": {
            "type": "object"
          },
          "responseError": {
            "$ref": "#/components/schemas/ResponseError"
          }
        }
      },
      "IPv4AddressMinimal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "ip": {
            "type": "string"
          },
          "netmask": {
            "type": "string"
          },
          "gateway": {
            "type": "string",
            "nullable": true
          },
          "broadcast": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ServerListMinimal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "hostname": {
            "type": "string",
            "nullable": true
          },
          "nickname": {
            "type": "string",
            "nullable": true
          },
          "disabled": {
            "type": "boolean"
          },
          "template": {
            "type": "object",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ServerTemplateMinimal"
              }
            ]
          }
        }
      },
      "ServerKeyboardLayoutPatch": {
        "type": "object",
        "properties": {
          "keyboardLayout": {
            "type": "string"
          }
        }
      },
...
      "TaskInfoStep": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TaskState"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2022-03-10T16:15:50Z"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2022-03-10T16:15:50Z"
          }
        }
      },
      "ServerState1": {
        "type": "string",
        "enum": [
          "ON",
          "OFF",
          "SUSPENDED"
        ]
      },
      "StorageOptimization": {
        "type": "string",
        "enum": [
          "INCONSISTENT",
          "COMPAT",
          "SLOW",
          "FAST",
          "NO"
        ]
      },
      "Bootorder": {
        "type": "string",
        "enum": [
          "HDD",
          "CDROM",
          "NETWORK"
        ]
      },
...
      "ServerInterface": {
        "type": "object",
        "properties": {
          "mac": {
            "type": "string"
          },
          "driver": {
            "type": "string"
          },
          "mtu": {
            "type": "integer",
            "format": "int32"
          },
          "speedInMBits": {
            "type": "integer",
            "format": "int32"
          },
          "rxMonthlyInMiB": {
            "type": "integer",
            "format": "int32"
          },
          "txMonthlyInMiB": {
            "type": "integer",
            "format": "int32"
          },
          "ipv4Addresses": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ipv6LinkLocalAddresses": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ipv6NetworkPrefixes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "trafficThrottled": {
            "type": "boolean"
          },
          "vlanInterface": {
            "type": "boolean"
          },
          "vlanId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Architecture": {
        "type": "string",
        "enum": [
          "AMD64",
          "ARM64"
        ]
      },
...
      "ServerHostnamePatch": {
        "type": "object",
        "properties": {
          "hostname": {
            "type": "string"
          }
        }
      },
...
      "ServerStatePatch": {
        "type": "object",
        "description": "Optional query parameters: stateOption.",
        "properties": {
          "state": {
            "$ref": "#/components/schemas/ServerState1"
          }
        }
      },
      "IPv6AddressMinimal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "networkPrefix": {
            "type": "string"
          },
          "networkPrefixLength": {
            "type": "integer",
            "format": "int32"
          },
          "gateway": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ServerTemplateMinimal": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 100,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9][a-zA-Z0-9 _,.\\/|()+-]{2,99}$",
            "type": "string"
          }
        }
      },
...
      "Site": {
        "required": [
          "name",
          "city"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "city": {
            "maxLength": 255,
            "pattern": "\\S",
            "type": "string"
          }
        }
      },
...
      "ServerInfo": {
        "type": "object",
        "properties": {
          "state": {
            "$ref": "#/components/schemas/ServerState"
          },
          "autostart": {
            "type": "boolean"
          },
          "uefi": {
            "type": "boolean"
          },
          "interfaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServerInterface"
            }
          },
          "disks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServerDisk"
            }
          },
          "bootorder": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Bootorder"
            }
          },
          "requiredStorageOptimization": {
            "$ref": "#/components/schemas/StorageOptimization"
          },
          "template": {
            "type": "string",
            "nullable": true
          },
          "uptimeInSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "currentServerMemoryInMiB": {
            "type": "integer",
            "format": "int64"
          },
          "maxServerMemoryInMiB": {
            "type": "integer",
            "format": "int64"
          },
          "cpuCount": {
            "type": "integer",
            "format": "int32"
          },
          "cpuMaxCount": {
            "type": "integer",
            "format": "int32"
          },
          "sockets": {
            "type": "integer",
            "format": "int32"
          },
          "coresPerSocket": {
            "type": "integer",
            "format": "int32"
          },
          "latestQemu": {
            "type": "boolean"
          },
          "configChanged": {
            "type": "boolean"
          },
          "osOptimization": {
            "$ref": "#/components/schemas/OsOptimization"
          },
          "nestedGuest": {
            "type": "boolean"
          },
          "machineType": {
            "type": "string"
          },
          "keyboardLayout": {
            "type": "string"
          },
          "cloudinitAttached": {
            "type": "boolean"
          }
        }
      },
...
      "ServerAutostartPatch": {
        "type": "object",
        "properties": {
          "autostart": {
            "type": "boolean"
          }
        }
      },
...
      "ServerNicknamePatch": {
        "type": "object",
        "properties": {
          "nickname": {
            "maxLength": 255,
            "type": "string"
          }
        }
      },
      "TaskProgress": {
        "type": "object",
        "properties": {
          "expectedFinishedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2022-03-10T16:15:50Z"
          },
          "progressInPercent": {
            "type": "number",
            "format": "float"
          }
        }
      },
...
      "ServerDisk": {
        "type": "object",
        "properties": {
          "dev": {
            "type": "string"
          },
          "driver": {
            "type": "string"
          },
          "capacityInMiB": {
            "type": "integer",
            "format": "int64"
          },
          "allocationInMiB": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ServerSetRootPasswordPatch": {
        "type": "object",
        "properties": {
          "rootPassword": {
            "type": "string"
          }
        }
      },
      "ServerBootorderPatch": {
        "required": [
          "bootorder"
        ],
        "type": "object",
        "properties": {
          "bootorder": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Bootorder"
            }
          }
        }
      },
...
      "ServerUEFIPatch": {
        "type": "object",
        "properties": {
          "uefi": {
            "type": "boolean"
          }
        }
      },
...
      "ServerState": {
        "type": "string",
        "enum": [
          "NOSTATE",
          "RUNNING",
          "BLOCKED",
          "PAUSED",
          "SHUTDOWN",
          "SHUTOFF",
          "CRASHED",
          "PMSUSPENDED",
          "DISK_SNAPSHOT"
        ]
      },
...
      "UserMinimal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "company": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "NotFoundError": {
        "type": "object",
        "example": {
          "message": "Entity Ticket with id 42 not found."
        },
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ServerOsOptimizationPatch": {
        "type": "object",
        "properties": {
          "os_optimization": {
            "$ref": "#/components/schemas/OsOptimization"
          }
        }
      },
      "CpuTopology": {
        "type": "object",
        "properties": {
          "socketCount": {
            "type": "integer",
            "format": "int32"
          },
          "coresPerSocketCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ServerCpuTopologyPatch": {
        "type": "object",
        "properties": {
          "cpuTopology": {
            "$ref": "#/components/schemas/CpuTopology"
          }
        }
      },
      "TaskState": {
        "type": "string",
        "enum": [
          "PENDING",
          "RUNNING",
          "FINISHED",
          "ERROR",
          "WAITING_FOR_CANCEL",
          "CANCELED",
          "ROLLBACK"
        ]
      },
...
    },
    "securitySchemes": {
      "Authorization": {
        "type": "openIdConnect",
        "description": "The netcup customer number originates from the Customer Control Panel (CCP). Server Control Panel (SCP) maintains a separate user base. The CCP customer number is used as the SCP username for authentication. The SCP userId is not equal to the CCP customer number; it is a unique internal identifier assigned by SCP.",
        "openIdConnectUrl": "/realms/scp/.well-known/openid-configuration"
      }
    }
  }
}